+++ src/lib/ebus/test/test_message.cpp
@@ -163,7 +163,7 @@ int main() {
   vector<SlaveSymbolString*> sstrs;
   mstrs.resize(1);
   sstrs.resize(1);
-  for (size_t i = 0; i < sizeof(checks) / sizeof(checks[0]); i++) {
+  for (unsigned int i = 0; i < sizeof(checks) / sizeof(checks[0]); i++) {
     string check[5] = checks[i];
     string inputStr = check[1];
     string flags = check[4];
